Move item, trigger, sector into elements
Also move picking into geometry
