+++ trview.app/Elements/Item.cpp
