+++ trview.app/Elements/Item.h
