+++ trview.app/Elements/RoomInfo.h
