+++ trview.app/Elements/Sector.cpp
