+++ trview.app/Elements/Sector.h
