+++ trview.app/Geometry/PickInfo.h
