+++ trview.app/Geometry/PickResult.cpp
