+++ trview.app/Geometry/PickResult.h
