+++ trview.app/Geometry/Picking.cpp
