+++ trview.app/Geometry/Picking.h
