+++ trview.app/Windows/ItemsWindow.h
@@ -4,7 +4,7 @@
 #pragma once
 
 #include <trview.ui/Listbox.h>
-#include <trview.app/Item.h>
+#include <trview.app/Elements/Item.h>
 #include "CollapsiblePanel.h"
 
 namespace trview
