+++ trview.app/trview.app.vcxproj
@@ -25,28 +25,28 @@
     <ClCompile Include="Camera\FreeCamera.cpp" />
     <ClCompile Include="Camera\ICamera.cpp" />
     <ClCompile Include="Camera\OrbitCamera.cpp" />
+    <ClCompile Include="Elements\Item.cpp" />
+    <ClCompile Include="Elements\Sector.cpp" />
+    <ClCompile Include="Elements\Trigger.cpp" />
     <ClCompile Include="Geometry\IRenderable.cpp" />
     <ClCompile Include="Geometry\Mesh.cpp" />
+    <ClCompile Include="Geometry\Picking.cpp" />
+    <ClCompile Include="Geometry\PickResult.cpp" />
     <ClCompile Include="Geometry\TransparencyBuffer.cpp" />
     <ClCompile Include="Geometry\TransparentTriangle.cpp" />
     <ClCompile Include="ILevelTextureStorage.cpp" />
-    <ClCompile Include="Item.cpp" />
     <ClCompile Include="ITextureStorage.cpp" />
     <ClCompile Include="Menus\DirectoryListing.cpp" />
     <ClCompile Include="Menus\FileDropper.cpp" />
     <ClCompile Include="Menus\LevelSwitcher.cpp" />
     <ClCompile Include="Menus\RecentFiles.cpp" />
     <ClCompile Include="Menus\ViewMenu.cpp" />
-    <ClCompile Include="Picking.cpp" />
-    <ClCompile Include="PickResult.cpp" />
     <ClCompile Include="Route.cpp" />
-    <ClCompile Include="Sector.cpp" />
     <ClCompile Include="SectorHighlight.cpp" />
     <ClCompile Include="SelectionRenderer.cpp" />
     <ClCompile Include="Tools\Compass.cpp" />
     <ClCompile Include="Tools\Measure.cpp" />
     <ClCompile Include="Tools\Toolbar.cpp" />
-    <ClCompile Include="Trigger.cpp" />
     <ClCompile Include="UI\CameraControls.cpp" />
     <ClCompile Include="UI\CameraPosition.cpp" />
     <ClCompile Include="UI\ContextMenu.cpp" />
@@ -75,33 +75,33 @@
     <ClInclude Include="Camera\FreeCamera.h" />
     <ClInclude Include="Camera\ICamera.h" />
     <ClInclude Include="Camera\OrbitCamera.h" />
+    <ClInclude Include="Elements\Item.h" />
+    <ClInclude Include="Elements\RoomInfo.h" />
+    <ClInclude Include="Elements\Sector.h" />
+    <ClInclude Include="Elements\Trigger.h" />
+    <ClInclude Include="Elements\Types.h" />
     <ClInclude Include="Geometry\IRenderable.h" />
     <ClInclude Include="Geometry\Mesh.h" />
     <ClInclude Include="Geometry\MeshVertex.h" />
+    <ClInclude Include="Geometry\PickInfo.h" />
+    <ClInclude Include="Geometry\Picking.h" />
+    <ClInclude Include="Geometry\PickResult.h" />
     <ClInclude Include="Geometry\TransparencyBuffer.h" />
     <ClInclude Include="Geometry\TransparentTriangle.h" />
     <ClInclude Include="Geometry\Triangle.h" />
     <ClInclude Include="ILevelTextureStorage.h" />
-    <ClInclude Include="Item.h" />
     <ClInclude Include="ITextureStorage.h" />
     <ClInclude Include="Menus\DirectoryListing.h" />
     <ClInclude Include="Menus\FileDropper.h" />
     <ClInclude Include="Menus\LevelSwitcher.h" />
     <ClInclude Include="Menus\RecentFiles.h" />
     <ClInclude Include="Menus\ViewMenu.h" />
-    <ClInclude Include="PickInfo.h" />
-    <ClInclude Include="Picking.h" />
-    <ClInclude Include="PickResult.h" />
-    <ClInclude Include="RoomInfo.h" />
     <ClInclude Include="Route.h" />
-    <ClInclude Include="Sector.h" />
     <ClInclude Include="SectorHighlight.h" />
     <ClInclude Include="SelectionRenderer.h" />
     <ClInclude Include="Tools\Compass.h" />
     <ClInclude Include="Tools\Measure.h" />
     <ClInclude Include="Tools\Toolbar.h" />
-    <ClInclude Include="Trigger.h" />
-    <ClInclude Include="Types.h" />
     <ClInclude Include="UI\CameraControls.h" />
     <ClInclude Include="UI\CameraPosition.h" />
     <ClInclude Include="UI\ContextMenu.h" />
