+++ trview.ui.render/MapRenderer.h
@@ -10,7 +10,7 @@
 #include <map>
 
 #include <trview.graphics/Sprite.h>
-#include <trview.app/Types.h>
+#include <trview.app/Elements/Types.h>
 #include "trview\TextureStorage.h"
 #include <trview.graphics/Texture.h>
 #include <trview.common/Point.h>
