+++ trview/Room.h
@@ -9,7 +9,7 @@
 #include <DirectXCollision.h>
 #include <SimpleMath.h>
 
-#include <trview.app/RoomInfo.h>
+#include <trview.app/Elements/RoomInfo.h>
 #include <trview.graphics/Texture.h>
 
 #include <trlevel/trtypes.h>
@@ -18,8 +18,8 @@
 #include "StaticMesh.h"
 #include <trview.app/Geometry/TransparencyBuffer.h>
 #include <trview.app/Geometry/Mesh.h>
-#include <trview.app/Sector.h>
-#include <trview.app/PickResult.h>
+#include <trview.app/Elements/Sector.h>
+#include <trview.app/Geometry/PickResult.h>
 
 namespace trview
 {
